ondongne_fullstack_ai_jwt_clip_yolo_clean/ai_service/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pandas==2.3.2
numpy==2.3.2
joblib==1.5.1
scikit-learn==1.7.1
python-multipart==0.0.20
pillow==11.3.0
transformers==4.55.4
torch==2.8.0
ultralytics==8.3.189
